refactor: rename api -> app and completely restructure/document it
