--- src/griptape_nodes/api/__init__.py
@@ -1,1 +1,0 @@
-"""App package."""
