--- src/griptape_nodes/api/queue_manager.py
@@ -1,3 +1,0 @@
-from queue import Queue
-
-event_queue = Queue()
