--- src/griptape_nodes/api/routes/__init__.py
@@ -1,1 +1,0 @@
-"""Flask routes package."""
