+++ src/griptape_nodes/app/.python-version
