+++ src/griptape_nodes/app/__init__.py
@@ -1,0 +1,5 @@
+"""App package."""
+
+from griptape_nodes.app.app import start_app
+
+__all__ = ["start_app"]
